chore: move kamelet repo test to advanced common
